Add Resampling Support
The new arguments are:

 (Sorry, there are no single character "short args" for these. We ran out of letters in the alphabet.)

	--sample-rate
		Sample Rate to Resample to {44.1kHz|48kHz|88.2kHz|96kHz}. Defaults to 44.1kHz meaning no resampling.

	--interpolation-quality
		Interpolation Quality to use if Resampling {Low|Medium|High}. Defaults to Low.

interpolation-quality: Low uses Linear interpolation.
interpolation-quality: Medium and High both use Windowed Sinc interpolation with a Blackman Window Function.

It goes without saying that the higher the sampling-rate and interpolation-quality the higher the CPU useage,
But I tried my best to optimize and balance preformance and audio quality as much as possible while still using only safe code.

The resampler is also dual threaded so that helps spread the load out across multi-core CPUs.

The end result is a pretty decent resampler that will run full tilt on basically a potato (my test device is a Pi Zero v2).